Exploring The SET Statement In SAS
A majority of DATA steps use the SET statement. The primary function of the SET statement in SAS is to read observations from one or more SAS datasets.

Properly using the SET statement in SAS is one of the key techniques for improving the efficiency of SAS programs. The SET statement has options that can be used to control how the data are to be read.
Read more at
https://www.9to5sas.com/set-statement-sas/