Create VarExists.sas
The %VAREXISTS macro uses Variable functions in a data step and to check if a variable exists in SAS and returns the variable info. It will return the below attributes of the variable when appropriate parameters are passed in the INFO parameter. If no value is passed, it returns the column position of the variable by default.

Len – To return the length of the variable.
fmt – to return Format of the variable
infmt – to return the informal of the variable.
label – to return Label of the variable
type – To return type of variable. (Character of Numeric).

Read more at - https://www.9to5sas.com/variable-exists-sas/